terminado el encontrar ciclos en un grafo de forma recursiva

+++ tp3/ej1/Adyacente.java
@@ -26,7 +26,7 @@ public class Adyacente {
 		return peso;
 	}
 
-	public void setPeso(int peso) {
+	public void setPeso(double peso) {
 		this.peso = peso;
 	}
 }
